Allow tile specials to passthru linalg to PXA
