+++ pmlc/conversion/linalg_to_pxa/pass_detail.h
@@ -10,6 +10,7 @@
 #include "mlir/Pass/Pass.h"
 #include "mlir/Transforms/DialectConversion.h"
 
+#include "pmlc/dialect/layer/ir/ops.h"
 #include "pmlc/dialect/pxa/ir/ops.h"
 #include "pmlc/dialect/stdx/ir/ops.h"
