+++ pmlc/conversion/tile_to_pxa/tile_to_pxa.cc
@@ -43,21 +43,6 @@ static Type getElementType(Value value) {
   return getElementType(value.getType());
 }
 
-static RankedTensorType getRankedTensorType(Type type) {
-  if (auto rankedTensorType = type.dyn_cast<RankedTensorType>()) {
-    return rankedTensorType;
-  }
-  return RankedTensorType::get({}, type);
-}
-
-static llvm::APFloat convertFloatUsingType(llvm::APFloat value,
-                                           FloatType type) {
-  bool losesInfo = false;
-  value.convert(type.getFloatSemantics(), APFloat::rmNearestTiesToEven,
-                &losesInfo);
-  return value;
-}
-
 static FlatSymbolRefAttr createStubTraceFunc(ModuleOp module, StringAttr msg) {
   static unsigned idCounter = 0;
   auto uniqueId = idCounter++;
@@ -377,48 +362,6 @@ struct LogicalNotOp {
   }
 };
 
-static Value createInit(OpBuilder &builder, Location loc, Type type,
-                        AggregationKind agg) {
-  if (auto floatType = type.dyn_cast<FloatType>()) {
-    switch (agg) {
-    case AggregationKind::add: {
-      auto value = convertFloatUsingType(llvm::APFloat(0.0), floatType);
-      return builder.create<mlir::ConstantFloatOp>(loc, value, floatType);
-    }
-    case AggregationKind::mul: {
-      auto value = convertFloatUsingType(llvm::APFloat(1.0), floatType);
-      return builder.create<mlir::ConstantFloatOp>(loc, value, floatType);
-    }
-    case AggregationKind::min: {
-      auto value = llvm::APFloat::getInf(floatType.getFloatSemantics(), false);
-      return builder.create<mlir::ConstantFloatOp>(loc, value, floatType);
-    }
-    case AggregationKind::max: {
-      auto value = llvm::APFloat::getInf(floatType.getFloatSemantics(), true);
-      return builder.create<mlir::ConstantFloatOp>(loc, value, floatType);
-    }
-    default:
-      llvm_unreachable("Unsupported aggregation for createInit");
-    }
-  } else if (auto intType = type.dyn_cast<IntegerType>()) {
-    switch (agg) {
-    case AggregationKind::add:
-      return builder.create<mlir::ConstantIntOp>(loc, 0, intType);
-    case AggregationKind::mul:
-      return builder.create<mlir::ConstantIntOp>(loc, 1, intType);
-    case AggregationKind::min:
-      return builder.create<mlir::ConstantIntOp>(
-          loc, std::numeric_limits<int>::max(), intType);
-    case AggregationKind::max:
-      return builder.create<mlir::ConstantIntOp>(
-          loc, std::numeric_limits<int>::min(), intType);
-    default:
-      llvm_unreachable("Unsupported aggregation for createInit");
-    }
-  }
-  llvm_unreachable("Unknown type for createInit");
-}
-
 template <typename CmpOpBuilder>
 struct CondOp {
   Value create(ConversionPatternRewriter &rewriter, Location loc,
@@ -432,35 +375,6 @@ struct CondOp {
         .getResult();
   }
 };
-
-static Value
-buildBroadcastLoad(OpBuilder &builder, Location loc, Value operand,
-                   unsigned outRank,
-                   Optional<tile::PaddingInfo> maybePadding = llvm::None) {
-  // Handle scalar values
-  if (!operand.getType().isa<MemRefType>()) {
-    return operand;
-  }
-  // handle broadcasts
-  auto body = builder.getBlock();
-  auto operandType = operand.getType().cast<MemRefType>();
-  assert(operandType.getRank() <= outRank && "result rank < operand rank");
-  ArrayRef<int64_t> shape = operandType.getShape();
-  SmallVector<Value, 8> operandIdxs(operandType.getRank());
-  for (unsigned i = 0; i < operandType.getRank(); i++) {
-    unsigned j = outRank - i - 1;
-    unsigned k = operandType.getRank() - i - 1;
-    if (shape[k] == 1) {
-      operandIdxs[k] = builder.create<mlir::ConstantIndexOp>(loc, 0);
-    } else {
-      operandIdxs[k] = body->getArgument(j);
-    }
-  }
-  auto loadOp = builder.create<pxa::PxaLoadOp>(loc, operand, operandIdxs);
-  if (maybePadding)
-    updateAffineMap(loadOp, *maybePadding);
-  return loadOp;
-}
 
 static AtomicRMWKind convertAgg(AggregationKind agg, Type type) {
   switch (agg) {
@@ -497,68 +411,6 @@ static AtomicRMWKind convertAgg(AggregationKind agg, Type type) {
   }
   llvm_unreachable("Invalid agg type in convertAgg");
 }
-
-struct BufferAllocator {
-  Value resultMemRef;
-  RankedTensorType rankedTensorType;
-  MemRefType memRefType;
-  Type elementType;
-
-  BufferAllocator(OpBuilder &builder, Operation *op, Type resultType) {
-    // Gather some basic info
-    TileToPXATypeConverter typeConverter;
-    auto loc = op->getLoc();
-    rankedTensorType = getRankedTensorType(resultType);
-    elementType = typeConverter.convertType(rankedTensorType.getElementType());
-    ArrayRef<int64_t> originalShape = rankedTensorType.getShape();
-    auto shape = llvm::to_vector<8>(originalShape);
-
-    // If padding is detected, expand the shape to accomodate.
-    auto maybePadding = tile::getPaddingInfo(op);
-    if (maybePadding) {
-      for (unsigned i = 0, e = shape.size(); i < e; ++i) {
-        shape[i] += maybePadding->lower[i] + maybePadding->upper[i];
-      }
-    }
-
-    // Make an allocation for the output
-    memRefType = MemRefType::get(shape, elementType);
-    resultMemRef = builder.create<memref::AllocOp>(loc, memRefType);
-    if (maybePadding) {
-      auto initValue = createInit(builder, loc, elementType, maybePadding->agg);
-      auto parallel = builder.create<AffineParallelOp>(
-          loc,
-          /*resultTypes=*/ArrayRef<Type>{memRefType},
-          /*reductions=*/ArrayRef<AtomicRMWKind>{AtomicRMWKind::assign},
-          /*ranges=*/shape);
-      auto parallelBuilder = parallel.getBodyBuilder();
-      auto load =
-          buildBroadcastLoad(parallelBuilder, loc, initValue, shape.size());
-      auto stored = buildSimpleStore(parallelBuilder, loc, load, resultMemRef,
-                                     llvm::None);
-      parallelBuilder.create<AffineYieldOp>(loc, ValueRange{stored});
-      resultMemRef = parallel.getResult(0);
-    }
-  }
-};
-
-struct PrngOpConversion : public OpConversionPattern<tile::PrngOp> {
-  using OpConversionPattern<tile::PrngOp>::OpConversionPattern;
-
-  LogicalResult
-  matchAndRewrite(tile::PrngOp op, ArrayRef<Value> operands,
-                  ConversionPatternRewriter &rewriter) const final {
-    tile::PrngOpAdaptor transformed(operands);
-    BufferAllocator allocResult(rewriter, op.getOperation(),
-                                op.result().getType());
-    BufferAllocator stateResult(rewriter, op.getOperation(),
-                                op.state().getType());
-    rewriter.replaceOpWithNewOp<pxa::PrngOp>(
-        op, allocResult.memRefType, stateResult.memRefType, transformed.state(),
-        allocResult.resultMemRef, stateResult.resultMemRef);
-    return success();
-  }
-};
 
 template <typename FromOpType, typename IntoOpBuilder,
           typename Matcher = AlwaysTrue>
@@ -1088,7 +940,6 @@ struct LowerTileToPXAPass : public LowerTileToPXABase<LowerTileToPXAPass> {
         FuncOpConversion<stdx::ClosureOp>, //
         IndexOpConversion,                 //
         PragmaOpConversion,                //
-        PrngOpConversion,                  //
         ReshapeOpConversion,               //
         ShapeOpConversion,                 //
         TraceOpConversion,                 //
