+++ pmlc/conversion/tile_to_pxa/utils.cc
@@ -10,6 +10,8 @@ namespace tile = dialect::tile;
 
 using namespace mlir; // NOLINT
 
+using util::AggregationKind;
+
 TileToPXATypeConverter::TileToPXATypeConverter() {
   addConversion([](FunctionType type) { return type; });
   addConversion([](FloatType type) { return type; });
@@ -108,6 +110,128 @@ void updateAffineMap(Operation *in, const tile::PaddingInfo &padding) {
   }
   accMap = AffineMap::get(accMap.getNumDims(), 0, newExprs, in->getContext());
   in->setAttr("map", AffineMapAttr::get(accMap));
+}
+
+static RankedTensorType getRankedTensorType(Type type) {
+  if (auto rankedTensorType = type.dyn_cast<RankedTensorType>()) {
+    return rankedTensorType;
+  }
+  return RankedTensorType::get({}, type);
+}
+
+llvm::APFloat convertFloatUsingType(llvm::APFloat value, FloatType type) {
+  bool losesInfo = false;
+  value.convert(type.getFloatSemantics(), APFloat::rmNearestTiesToEven,
+                &losesInfo);
+  return value;
+}
+
+Value createInit(OpBuilder &builder, Location loc, Type type,
+                 AggregationKind agg) {
+  if (auto floatType = type.dyn_cast<FloatType>()) {
+    switch (agg) {
+    case AggregationKind::add: {
+      auto value = convertFloatUsingType(llvm::APFloat(0.0), floatType);
+      return builder.create<mlir::ConstantFloatOp>(loc, value, floatType);
+    }
+    case AggregationKind::mul: {
+      auto value = convertFloatUsingType(llvm::APFloat(1.0), floatType);
+      return builder.create<mlir::ConstantFloatOp>(loc, value, floatType);
+    }
+    case AggregationKind::min: {
+      auto value = llvm::APFloat::getInf(floatType.getFloatSemantics(), false);
+      return builder.create<mlir::ConstantFloatOp>(loc, value, floatType);
+    }
+    case AggregationKind::max: {
+      auto value = llvm::APFloat::getInf(floatType.getFloatSemantics(), true);
+      return builder.create<mlir::ConstantFloatOp>(loc, value, floatType);
+    }
+    default:
+      llvm_unreachable("Unsupported aggregation for createInit");
+    }
+  } else if (auto intType = type.dyn_cast<IntegerType>()) {
+    switch (agg) {
+    case AggregationKind::add:
+      return builder.create<mlir::ConstantIntOp>(loc, 0, intType);
+    case AggregationKind::mul:
+      return builder.create<mlir::ConstantIntOp>(loc, 1, intType);
+    case AggregationKind::min:
+      return builder.create<mlir::ConstantIntOp>(
+          loc, std::numeric_limits<int>::max(), intType);
+    case AggregationKind::max:
+      return builder.create<mlir::ConstantIntOp>(
+          loc, std::numeric_limits<int>::min(), intType);
+    default:
+      llvm_unreachable("Unsupported aggregation for createInit");
+    }
+  }
+  llvm_unreachable("Unknown type for createInit");
+}
+
+Value buildBroadcastLoad(OpBuilder &builder, Location loc, Value operand,
+                         unsigned outRank,
+                         Optional<tile::PaddingInfo> maybePadding) {
+  // Handle scalar values
+  if (!operand.getType().isa<MemRefType>()) {
+    return operand;
+  }
+  // handle broadcasts
+  auto body = builder.getBlock();
+  auto operandType = operand.getType().cast<MemRefType>();
+  assert(operandType.getRank() <= outRank && "result rank < operand rank");
+  ArrayRef<int64_t> shape = operandType.getShape();
+  SmallVector<Value, 8> operandIdxs(operandType.getRank());
+  for (unsigned i = 0; i < operandType.getRank(); i++) {
+    unsigned j = outRank - i - 1;
+    unsigned k = operandType.getRank() - i - 1;
+    if (shape[k] == 1) {
+      operandIdxs[k] = builder.create<mlir::ConstantIndexOp>(loc, 0);
+    } else {
+      operandIdxs[k] = body->getArgument(j);
+    }
+  }
+  auto loadOp = builder.create<pxa::PxaLoadOp>(loc, operand, operandIdxs);
+  if (maybePadding)
+    updateAffineMap(loadOp, *maybePadding);
+  return loadOp;
+}
+
+BufferAllocator::BufferAllocator(OpBuilder &builder, Operation *op,
+                                 Type resultType) {
+  // Gather some basic info
+  TileToPXATypeConverter typeConverter;
+  auto loc = op->getLoc();
+  rankedTensorType = getRankedTensorType(resultType);
+  elementType = typeConverter.convertType(rankedTensorType.getElementType());
+  ArrayRef<int64_t> originalShape = rankedTensorType.getShape();
+  auto shape = llvm::to_vector<8>(originalShape);
+
+  // If padding is detected, expand the shape to accomodate.
+  auto maybePadding = tile::getPaddingInfo(op);
+  if (maybePadding) {
+    for (unsigned i = 0, e = shape.size(); i < e; ++i) {
+      shape[i] += maybePadding->lower[i] + maybePadding->upper[i];
+    }
+  }
+
+  // Make an allocation for the output
+  memRefType = MemRefType::get(shape, elementType);
+  resultMemRef = builder.create<memref::AllocOp>(loc, memRefType);
+  if (maybePadding) {
+    auto initValue = createInit(builder, loc, elementType, maybePadding->agg);
+    auto parallel = builder.create<AffineParallelOp>(
+        loc,
+        /*resultTypes=*/ArrayRef<Type>{memRefType},
+        /*reductions=*/ArrayRef<AtomicRMWKind>{AtomicRMWKind::assign},
+        /*ranges=*/shape);
+    auto parallelBuilder = parallel.getBodyBuilder();
+    auto load =
+        buildBroadcastLoad(parallelBuilder, loc, initValue, shape.size());
+    auto stored =
+        buildSimpleStore(parallelBuilder, loc, load, resultMemRef, llvm::None);
+    parallelBuilder.create<AffineYieldOp>(loc, ValueRange{stored});
+    resultMemRef = parallel.getResult(0);
+  }
 }
 
 } // namespace pmlc::conversion::tile_to_pxa
