+++ pmlc/dialect/tile/ir/ops.td
@@ -96,10 +96,10 @@ class MaterializeOp<string mnemonic, list<OpTrait> traits = [NoSideEffect]> :
 def ArgSortOp : MaterializeOp<"argsort"> {
   let summary = "tensor sort index operation";
   let arguments = (ins
-    RankedTensorOf<[AnyScalar]>:$tensor,
+    RankedTensorOf<[AnyInteger, AnyScalar]>:$tensor,
     IndexAttr:$axis,
     SortDirection:$direction);
-  let results = (outs RankedTensorOf<[SI32]>:$result);
+  let results = (outs RankedTensorOf<[I32, SI32]>:$result);
   let assemblyFormat = [{
     $direction $tensor `[` $axis `]` attr-dict `:` functional-type(operands, $result)
   }];
@@ -108,8 +108,8 @@ def ArgSortOp : MaterializeOp<"argsort"> {
 def GatherOp : MaterializeOp<"gather"> {
   let summary = "special gather operation";
   let arguments = (ins
-    RankedTensorOf<[AnyScalar]>:$tensor,
-    RankedTensorOf<[AnyScalar]>:$indices,
+    RankedTensorOf<[AnyInteger, AnyScalar]>:$tensor,
+    RankedTensorOf<[AnyInteger, AnyScalar]>:$indices,
     DefaultValuedAttr<IndexAttr, "0">:$axis,
     DefaultValuedAttr<InterpolationMode, "InterpolationMode::linear">:$interpolationMode,
     DefaultValuedAttr<NearestMode, "NearestMode::round_prefer_floor">:$nearestMode,
@@ -118,7 +118,7 @@ def GatherOp : MaterializeOp<"gather"> {
     DefaultValuedAttr<IndexAttr, "0">:$batchDims,
     DefaultValuedAttr<OutOfBoundsMode, "OutOfBoundsMode::gather_edge_padded_input">:$OutOfBoundsMode
   );
-  let results = (outs RankedTensorOf<[AnyScalar]>:$result);
+  let results = (outs RankedTensorOf<[AnyInteger, AnyScalar]>:$result);
   let assemblyFormat = [{
     $tensor $indices attr-dict `:` functional-type(operands, $result)
   }];
@@ -145,12 +145,12 @@ def IndexOp : SpecialOp<"index"> {
   let assemblyFormat = "$axis attr-dict `:` type($result)";
 }
 
-def PrngOp : SpecialOp<"prng"> {
+def PrngOp : SpecialOp<"prng", []> {
   let summary = "pseudorandom number generator";
-  let arguments = (ins RankedTensorOf<[UI32]>:$state);
+  let arguments = (ins RankedTensorOf<[I32, UI32]>:$state);
   let results = (outs
     RankedTensorOf<[AnyFloat]>:$result,
-    RankedTensorOf<[UI32]>:$new_state);
+    RankedTensorOf<[I32, UI32]>:$new_state);
   let assemblyFormat = "$state attr-dict `:` functional-type($state, results)";
 }
 
@@ -168,13 +168,13 @@ def ReshapeOp : MaterializeOp<"reshape">,
 def ScatterOp : MaterializeOp<"scatter"> {
   let summary = "special scatter operation";
   let arguments = (ins
-    RankedTensorOf<[AnyScalar]>:$data,
-    RankedTensorOf<[AnyScalar]>:$indices,
-    RankedTensorOf<[AnyScalar]>:$updates,
+    RankedTensorOf<[AnyInteger, AnyScalar]>:$data,
+    RankedTensorOf<[AnyInteger, AnyScalar]>:$indices,
+    RankedTensorOf<[AnyInteger, AnyScalar]>:$updates,
     DefaultValuedAttr<IndexAttr, "0">:$axis,
     DefaultValuedAttr<ScatterMode, "ScatterMode::normal">:$mode
   );
-  let results = (outs RankedTensorOf<[AnyScalar]>:$result);
+  let results = (outs RankedTensorOf<[AnyInteger, AnyScalar]>:$result);
   let builders = [
     OpBuilder<(ins
       "mlir::Type":$resultType,
